Move ssr webpack configs to its own file
